feat(hexagon): P3a galil split slice-1 — lift TransformXY to hexagon domain service (D6, P4-shared)
Moves TransformXY verbatim out of galil_motion_driver.py into
helao/hexagon/domain/motion_transform.py: a pure motor/plate/instrument
coordinate-transform service with no Base/Active/gclib/Bokeh/file-IO
coupling, now Linux-testable and reusable by P4's ThorlabsMotor. The
driver imports the class from its new location; the stray debug
print() in transform_motorxy_to_platexy is preserved for parity with a
TODO(P3a) marker. LOGGER uses stdlib logging.getLogger(__name__) per
the domain allow-list (matches global_params.py/dispatch_policy.py
convention), not helao.helpers.helao_logging.
